refactor file structure. add dockerfile
