+++ historical_data_collectors/historical_runner.py
@@ -1,17 +1,9 @@
-import requests
 import sys
-# import psycopg2
-import ccxt
 import datetime
-import pytz
 import time
-# import paramiko
-# from sshtunnel import SSHTunnelForwarder
-# from paramiko.rsakey import RSAKey
 
-from historical_data_collectors.binance_data_collector import BinanceDataCollector
-from historical_data_collectors.coinbase_data_collector import CoinbaseDataCollector
-
+from historical_data_collectors.collectors.binance_data_collector import BinanceDataCollector
+from historical_data_collectors.collectors.coinbase_data_collector import CoinbaseDataCollector
 
 def main():
 
@@ -21,6 +13,19 @@ def main():
         if len(sys.argv) != 4:
             print("Usage: python3 historical_runner.py exchange_name start_date end_date")
             sys.exit(1)
+
+        arg_date_format = "%Y/%m/%d"
+
+        #inclusive
+        start_date = datetime.datetime.strptime(sys.argv[2], arg_date_format).date()
+
+        #exclusive
+        end_date = datetime.datetime.strptime(sys.argv[3], arg_date_format).date()
+
+        if start_date >= end_date:
+            print("Start date needs to be before end date, no data fetched")
+            sys.exit(1)
+        
 
         exchange_name = sys.argv[1].lower()
 
@@ -40,14 +45,6 @@ def main():
             print(f"Exchange {exchange_name} is not supported. Currently supported exchanges are Binance, Coinbase, Dydx, Bybit and Okx")
             sys.exit(1)
 
-        arg_date_format = "%Y/%m/%d"
-
-        #inclusive
-        start_date = datetime.datetime.strptime(sys.argv[2], arg_date_format).date()
-
-        #exclusive
-        end_date = datetime.datetime.strptime(sys.argv[3], arg_date_format).date()
-
         
         data_collector.fetch_and_write_trades(start_date, end_date)
     
@@ -58,8 +55,6 @@ def main():
         execution_time = end_time - start_time
 
         print("Script execution time: {:.2f} seconds".format(execution_time))
-
-    
 
 if __name__ == "__main__":
     main()
